コンポーネントの配置を修正
