+++ src/pages/components/Closet/registration_cloth/ClothingRegistration.tsx
