+++ src/pages/dashboard/register-clothes.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Layout from '../components/Closet//Layout';
-import ClothingRegistration from '../components/Closet/ClothingRegistration';
+import Layout from '../components/closet/Layout';
+import ClothingRegistration from '../components/closet/registration_cloth/ClothingRegistration';
 
 const RegisterClothesPage: React.FC = () => {
   return (
