+++ src/pages/dashboard/view-closet.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Layout from '../components/Closet/Layout';
-import ClothesList from '../components/Closet/ClothesList';
+import Layout from '../components/closet/Layout';
+import ClothesList from '../components/closet/clotheslist/ClothesList';
 
 const ViewClosetPage: React.FC = () => {
   return (
